weather_usecase_dag_1/requirements.txt
mysql-connector-python
numpy
pandas
python-dotenv
pytz
requests
openmeteo-requests
requests-cache 
retry-requests